# v-sign-be/openapi-backend.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
    "/api/v1/lessons/{lessonId}/progress" : {
      "put" : {
        "tags" : [ "learning-catalog-controller" ],
        "operationId" : "updateProgress",
        "parameters" : [ {
          "name" : "lessonId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateProgressRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponseProgressResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/payments/orders" : {
      "post" : {
        "tags" : [ "payment-controller" ],
        "operationId" : "createOrder",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreatePaymentOrderRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponsePaymentOrderResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/payments/{transactionId}" : {
      "patch" : {
        "tags" : [ "admin-payment-controller" ],
        "operationId" : "overrideStatus",
        "parameters" : [ {
          "name" : "transactionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ManualPaymentStatusRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponseAdminPaymentRecordResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/units" : {
      "get" : {
        "tags" : [ "learning-catalog-controller" ],
        "operationId" : "listUnits",
        "parameters" : [ {
          "name" : "publishedOnly",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : true
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "maximum" : 100,
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponseUnitListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/units/{unitId}/chapters" : {
      "get" : {
        "tags" : [ "learning-catalog-controller" ],
        "operationId" : "listChapters",
        "parameters" : [ {
          "name" : "unitId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponseChapterListResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/payments/{transactionId}" : {
      "get" : {
        "tags" : [ "payment-controller" ],
        "operationId" : "status",
        "parameters" : [ {
          "name" : "transactionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponsePaymentStatusResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/lessons/{lessonId}" : {
      "get" : {
        "tags" : [ "learning-catalog-controller" ],
        "operationId" : "getLesson",
        "parameters" : [ {
          "name" : "lessonId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponseLessonDetailResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/learning/practice-items" : {
      "get" : {
        "tags" : [ "learning-controller" ],
        "operationId" : "listPracticeItems",
        "parameters" : [ {
          "name" : "category",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "level",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "maximum" : 100,
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponsePracticeItemsPageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/learning/practice-items/{itemId}" : {
      "get" : {
        "tags" : [ "learning-controller" ],
        "operationId" : "getPracticeItem",
        "parameters" : [ {
          "name" : "itemId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponsePracticeItemDetailResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leaderboards" : {
      "get" : {
        "tags" : [ "leaderboard-controller" ],
        "operationId" : "leaderboard",
        "parameters" : [ {
          "name" : "period",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "WEEKLY",
            "enum" : [ "WEEKLY", "MONTHLY" ]
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "maximum" : 50,
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponseLeaderboardResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/dictionary" : {
      "get" : {
        "tags" : [ "dictionary-public-controller" ],
        "operationId" : "list",
        "parameters" : [ {
          "name" : "category",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "keyword",
          "in" : "query",
          "required" : false,
          "schema" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          }
        }, {
          "name" : "difficulty",
          "in" : "query",
          "required" : false,
          "schema" : {
            "maxLength" : 20,
            "minLength" : 0,
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "maximum" : 100,
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponseDictionaryEntryPage"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dictionary/{entryId}/practice-target" : {
      "get" : {
        "tags" : [ "dictionary-public-controller" ],
        "operationId" : "practiceTarget",
        "parameters" : [ {
          "name" : "entryId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponsePracticeTargetResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chapters/{chapterId}/lessons" : {
      "get" : {
        "tags" : [ "learning-catalog-controller" ],
        "operationId" : "listLessons",
        "parameters" : [ {
          "name" : "chapterId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SuccessResponseLessonListResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "UpdateProgressRequest" : {
        "type" : "object",
        "properties" : {
          "completionPct" : {
            "maximum" : 100,
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          },
          "lastPositionSeconds" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          },
          "phase" : {
            "type" : "string"
          },
          "currentQuestionIndex" : {
            "type" : "integer",
            "format" : "int32"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "NOT_STARTED", "IN_PROGRESS", "COMPLETED" ]
          }
        }
      },
      "ProgressResponse" : {
        "type" : "object",
        "properties" : {
          "lessonId" : {
            "type" : "string"
          },
          "completionPct" : {
            "type" : "integer",
            "format" : "int32"
          },
          "lastPositionSeconds" : {
            "type" : "integer",
            "format" : "int32"
          },
          "phase" : {
            "type" : "string"
          },
          "currentQuestionIndex" : {
            "type" : "integer",
            "format" : "int32"
          },
          "status" : {
            "type" : "string"
          }
        }
      },
      "SuccessResponseProgressResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/ProgressResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "CreatePaymentOrderRequest" : {
        "required" : [ "provider" ],
        "type" : "object",
        "properties" : {
          "provider" : {
            "pattern" : "MOMO|ZALOPAY",
            "type" : "string"
          },
          "planId" : {
            "type" : "string"
          },
          "planType" : {
            "pattern" : "MONTHLY|YEARLY",
            "type" : "string"
          },
          "amount" : {
            "minimum" : 49000,
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "PaymentOrderResponse" : {
        "type" : "object",
        "properties" : {
          "transactionId" : {
            "type" : "string"
          },
          "providerTransactionId" : {
            "type" : "string"
          },
          "provider" : {
            "type" : "string"
          },
          "planId" : {
            "type" : "string"
          },
          "planType" : {
            "type" : "string"
          },
          "amount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "currency" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "qrCodeData" : {
            "type" : "string"
          },
          "deepLink" : {
            "type" : "string"
          },
          "expiresAt" : {
            "type" : "string"
          },
          "qrCodeUrl" : {
            "type" : "string"
          },
          "expiresInSeconds" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "SuccessResponsePaymentOrderResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PaymentOrderResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "ManualPaymentStatusRequest" : {
        "required" : [ "reason", "status" ],
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string"
          },
          "reason" : {
            "maxLength" : 500,
            "minLength" : 10,
            "type" : "string"
          }
        }
      },
      "AdminPaymentRecordResponse" : {
        "type" : "object",
        "properties" : {
          "transactionId" : {
            "type" : "string"
          },
          "userEmail" : {
            "type" : "string"
          },
          "planId" : {
            "type" : "string"
          },
          "amount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "currency" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "provider" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string"
          },
          "updatedAt" : {
            "type" : "string"
          },
          "overrideReason" : {
            "type" : "string"
          }
        }
      },
      "SuccessResponseAdminPaymentRecordResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/AdminPaymentRecordResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "SuccessResponseUnitListResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/UnitListResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "UnitListResponse" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "units" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UnitSummaryResponse"
            }
          }
        }
      },
...
      "ChapterListResponse" : {
        "type" : "object",
        "properties" : {
          "unitId" : {
            "type" : "string"
          },
          "chapters" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ChapterSummaryResponse"
            }
          }
        }
      },
...
      "SuccessResponseChapterListResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/ChapterListResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "PaymentStatusResponse" : {
        "type" : "object",
        "properties" : {
          "transactionId" : {
            "type" : "string"
          },
          "providerTransactionId" : {
            "type" : "string"
          },
          "provider" : {
            "type" : "string"
          },
          "planType" : {
            "type" : "string"
          },
          "amount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "currency" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string"
          },
          "retryable" : {
            "type" : "boolean"
          }
        }
      },
      "SuccessResponsePaymentStatusResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PaymentStatusResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "LessonDetailResponse" : {
        "type" : "object",
        "properties" : {
          "lessonId" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "videoUrl" : {
            "type" : "string"
          },
          "requiresPremium" : {
            "type" : "boolean"
          },
          "progress" : {
            "$ref" : "#/components/schemas/LessonProgressCheckpointResponse"
          }
        }
      },
...
      "SuccessResponseLessonDetailResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/LessonDetailResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "PracticeItemsPageResponse" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PracticeItemSummaryResponse"
            }
          }
        }
      },
      "SuccessResponsePracticeItemsPageResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PracticeItemsPageResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "SuccessResponsePracticeItemDetailResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PracticeItemDetailResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "SuccessResponseLeaderboardResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/LeaderboardResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "SuccessResponseDictionaryEntryPage" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/DictionaryEntryPage"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "SuccessResponsePracticeTargetResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PracticeTargetResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "SuccessResponseLessonListResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/LessonListResponse"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
    }
  }
}